+++ lib/c++/src/cave_talk.cc
@@ -359,7 +359,7 @@ CaveTalk_Error_t Listener::HandleAirQuality(CaveTalk_Length_t length) const
     }
 
     const uint32_t dust_ug_per_m3 = air_quality_message.dust_ug_per_m3();
-    const double gas_ppm = air_quality_message.gas_ppm();
+    const uint32_t gas_ppm        = air_quality_message.gas_ppm();
 
     listener_callbacks_->HearAirQuality(dust_ug_per_m3, gas_ppm);
 
@@ -555,7 +555,7 @@ CaveTalk_Error_t Talker::SpeakConfigSteeringControl(const PID &turn_rate_params,
     return CaveTalk_Speak(&link_handle_, static_cast<CaveTalk_Id_t>(ID_CONFIG_STEERING_CONTROL), message_buffer_.data(), length);
 }
 
-CaveTalk_Error_t Talker::SpeakAirQuality(const uint32_t dust_ug_per_m3, const double gas_ppm)
+CaveTalk_Error_t Talker::SpeakAirQuality(const uint32_t dust_ug_per_m3, const uint32_t gas_ppm)
 {
     AirQuality air_quality_message;
     air_quality_message.set_dust_ug_per_m3(dust_ug_per_m3);
